requirements.txt
pandas
Pillow
numpy
scikit-learn
